refactor(modal): move modal-ref and modal-config in separate files
