+++ src/framework/theme/components/modal/index.ts
@@ -1,0 +1,4 @@
+export * from './modal-config';
+export * from './modal-ref';
+export * from './modal.service';
+export * from './modal.module';
