+++ src/framework/theme/index.ts
@@ -71,5 +71,4 @@ export * from './components/list/infinite-list.directive';
 export * from './components/input/input.directive';
 export * from './components/input/input.module';
 export * from './components/cdk/overlay';
-export * from './components/modal/modal.module';
-export * from './components/modal/modal.service';
+export * from './components/modal';
